requirements.txt
# empEnglish Requirements
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.1.0
pymongo==4.6.1
redis==5.0.1
pymilvus==2.3.4

# AI/ML
langchain==0.1.0
langgraph==0.0.26
openai==1.6.1
faster-whisper==1.0.0
edge-tts==6.1.12

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# Security
cryptography==41.0.8
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3.post1

# Performance Testing
locust==2.19.2

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
